If using -command-stop don't restart on build fail
When running a web serivce it is nice to know if the build breaks by
stopping the service. The -command-stop seems like it should make this
use case work, but it restarted the command regards of the success of
the build. This fix does not restart the process if the build failed.