Add dep versioning via `go mod`
This change pins direct dependencies to specific versions, and enables
developing and building this project outside of `$GOPATH`.

This change was automatically generated via:

```
$ go mod init github.com/githubnemo/CompileDaemon
$ go mod tidy
```

Since this feature was introduced in Go 1.11, this is now the minimum Go version
to compile this project; however, note that since the Go project only maintains
the 2 most recent major releases [1], and the two most recent releases are Go
1.13 and 1.14, any releases up to and including Go 1.12 are no longer supported
anyway.

This change also requires setting the env variable `GO111MODULE=on` to work.

[1] https://golang.org/doc/devel/release.html says:

> ## Release Policy
>
> Each major Go release is supported until there are two newer major releases.
> For example, Go 1.5 was supported until the Go 1.7 release, and Go 1.6 was
> supported until the Go 1.8 release. [...]
>
> ## go1.14 (released 2020/02/25)